requirements/prod.txt
asgiref==3.2.7
Django==3.0.5
django-filter==21.1
django-model-utils==4.0.0
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1
